+++ src/main/kotlin/app/revanced/patches/youtube/misc/minimizedplayback/patch/MinimizedPlaybackPatch.kt
@@ -5,6 +5,7 @@ import app.revanced.patcher.BytecodeContext
 import app.revanced.patcher.annotation.Description
 import app.revanced.patcher.annotation.Name
 import app.revanced.patcher.annotation.Version
+import app.revanced.patcher.extensions.addInstruction
 import app.revanced.patcher.extensions.addInstructions
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchResult
@@ -15,6 +16,7 @@ import app.revanced.patches.shared.settings.preference.impl.NonInteractivePrefer
 import app.revanced.patches.shared.settings.preference.impl.StringResource
 import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
 import app.revanced.patches.youtube.misc.minimizedplayback.annotations.MinimizedPlaybackCompatibility
+import app.revanced.patches.youtube.misc.minimizedplayback.fingerprints.KidsMinimizedPlaybackPolicyControllerFingerprint
 import app.revanced.patches.youtube.misc.minimizedplayback.fingerprints.MinimizedPlaybackManagerFingerprint
 import app.revanced.patches.youtube.misc.minimizedplayback.fingerprints.MinimizedPlaybackSettingsFingerprint
 import app.revanced.patches.youtube.misc.playertype.patch.PlayerTypeHookPatch
@@ -31,7 +33,8 @@ import org.jf.dexlib2.iface.reference.MethodReference
 class MinimizedPlaybackPatch : BytecodePatch(
     listOf(
         MinimizedPlaybackManagerFingerprint,
-        MinimizedPlaybackSettingsFingerprint
+        MinimizedPlaybackSettingsFingerprint,
+        KidsMinimizedPlaybackPolicyControllerFingerprint
     )
 ) {
     override fun execute(context: BytecodeContext): PatchResult {
@@ -45,7 +48,7 @@ class MinimizedPlaybackPatch : BytecodePatch(
         MinimizedPlaybackManagerFingerprint.result?.apply {
             mutableMethod.addInstructions(
                 0, """
-                invoke-static {}, $INTEGRATIONS_CLASS_DESCRIPTOR->playbackIsNotShort()Z
+                invoke-static {}, $INTEGRATIONS_CLASS_DESCRIPTOR->isPlaybackNotShort()Z
                 move-result v0
                 return v0
                 """
@@ -69,6 +72,15 @@ class MinimizedPlaybackPatch : BytecodePatch(
                 """
             )
         } ?: return MinimizedPlaybackSettingsFingerprint.toErrorResult()
+
+        // Force allowing background play for videos labeled for kids.
+        // Some regions and YouTube accounts do not require this patch.
+        KidsMinimizedPlaybackPolicyControllerFingerprint.result?.apply {
+            mutableMethod.addInstruction(
+                0,
+                "return-void"
+            )
+        } ?: return KidsMinimizedPlaybackPolicyControllerFingerprint.toErrorResult()
 
         return PatchResult.Success
     }